Large file support!
https://github.com/fogWraith/Hotline/blob/main/Docs/Protocol/Capabilities-Large-File.md

+++ hotline-tauri/src-tauri/src/commands/mod.rs
@@ -162,7 +162,7 @@ pub async fn download_file(
     server_id: String,
     path: Vec<String>,
     file_name: String,
-    file_size: u32,
+    file_size: u64,
     download_folder: Option<String>,
     state: State<'_, AppState>,
 ) -> Result<String, String> {
